Slight refactor to have less of an overhead
