+++ go.mod
@@ -3,9 +3,12 @@ module github.com/benmatselby/lionel
 require (
 	github.com/benmatselby/walter v0.0.0-20190124111714-a73de48e5bd1
 	github.com/golang/mock v1.2.0
+	github.com/google/pprof v0.0.0-20190109223431-e84dfd68c163 // indirect
+	github.com/ianlancetaylor/demangle v0.0.0-20181102032728-5e5cf60278f6 // indirect
 	github.com/mitchellh/go-homedir v1.0.0
 	github.com/spf13/afero v1.2.0 // indirect
 	github.com/spf13/cobra v0.0.3
 	github.com/spf13/viper v1.3.1
+	golang.org/x/arch v0.0.0-20181203225421-5a4828bb7045 // indirect
 	golang.org/x/sys v0.0.0-20181217223516-dcdaa6325bcb // indirect
 )
